refactor: split oneclick service in inscription service and authorization service
